Bugs fixed and security things

+++ tools_and_games/PT_games/Photo_Hunt/core/src/com/potatoandtomato/games/screens/main/MainLogic.java
@@ -282,7 +282,7 @@ public class MainLogic extends GameLogic {
             _gameModel.setGameState(GameState.Pause);
             if(meIsThisStageDecisionMaker()){
                 _services.getRoomMsgHandler().sendWon(new WonStageModel(_gameModel.getStageNumber(),
-                                        _gameModel.getScore(), remainingSecs, _gameModel.getHintsLeft(), "0", StageType.Normal));
+                                        _gameModel.getScore().intValue(), remainingSecs, _gameModel.getHintsLeft(), "0", StageType.Normal));
             }
         }
     }
